edition in forumcontroller

+++ ScarboroughFair/src/main/java/icu/internetcomputation/scarboroughfair/service/ForumService.java
@@ -40,27 +40,38 @@ public class ForumService {
         return commentRepository.findById(id).orElse(null);
     }
 
-    public Message addPost(User user, String content, String[] imgUrl){
+    public Message addPost(int userid, String content, String[] imgUrl){
         int postid = (int) (forumPostRepository.count() + 1);
-        int userid = user.getId();
         Long time = System.currentTimeMillis();
         ForumPost post = new ForumPost(postid, userid, content, imgUrl, time);
         forumPostRepository.save(post);
         return new Message(true, "帖子发布成功o(￣ε￣*)", null, postid);
     }
 
-    public Message deletePost(int postid){
+    public Message deletePost(int userid, int postid){
         ForumPost delpost= findForumPost(postid);
+        int trueuser = delpost.getPostUserID();
+        if(userid != trueuser){
+            return new Message(false, "帖子被神秘的东方力量阻挡，无法删除");
+        }
         forumPostRepository.delete(delpost);
         return new Message(true, "帖子忍痛离开了普罗丢瑟sama~( TロT)");
     }
 
-    public Message addComment(User user, String content, Integer postId){
+    public Message addComment(int userid, String content, Integer postId){
         int commentid = (int) (commentRepository.count() + 1);
-        int userid = user.getId();
         Long time = System.currentTimeMillis();
         Comment comment = new Comment(postId, commentid, userid, time, content);
         commentRepository.save(comment);
+        ForumPost forumPost = findForumPost(postId);
+        int[] oldCommentID = forumPost.getCommentID();
+        int[] newCommentID = new int[oldCommentID.length + 1];
+        for(int i = 0; i < oldCommentID.length; i++){
+            newCommentID[i] = oldCommentID[i];
+        }
+        newCommentID[oldCommentID.length] = commentid;
+        forumPost.setCommentID(newCommentID);
+        forumPostRepository.save(forumPost);
         return new Message(true, "您的评论已经传达到辣d=====(￣▽￣*)b");
     }
 
@@ -71,9 +82,26 @@ public class ForumService {
         return new Message(true, "非常好内容, 爱来自陶瓷");
     }
 
-    public Message deleteComment(Integer commentid){
+    public Message deleteComment(int userid, int commentid){
         Comment delComment = findComment(commentid);
+        if(userid != delComment.getCommentUserID()){
+            return new Message(false, "删除评论的权力被我夺舍了!");
+        }
+        int delCommentId = delComment.getCommentId();
+        ForumPost forumPost = findForumPost(delComment.getPostId());
+        int[] oldCommentID = forumPost.getCommentID();
+        int[] newCommentID = new int[oldCommentID.length - 1];
+        int i = 0, j = 0;
+        while(i != oldCommentID.length - 1){
+            if(oldCommentID[j] != delCommentId){
+                newCommentID[i] = oldCommentID[j];
+                i++;
+            }
+            j++;
+        }
+        forumPost.setCommentID(newCommentID);
+        forumPostRepository.save(forumPost);
         commentRepository.delete(delComment);
-        return new Message(true, "评论帮你删好了，感谢我吧!");
+        return new Message(true, "评论帮你删好了，再多夸夸我也没关系哦!");
     }
 }
